Move validation policies to Stratis.SmartContracts.CLR
